+++ src/Symfony/Component/ErrorRenderer/Exception/FlattenException.php
@@ -22,10 +22,8 @@ use Symfony\Component\HttpKernel\Exception\HttpExceptionInterface;
  * Basically, this class removes all objects from the trace.
  *
  * @author Fabien Potencier <fabien@symfony.com>
- *
- * @internal
  */
-class FlattenException
+final class FlattenException
 {
     private $title;
     private $message;
@@ -39,7 +37,7 @@ class FlattenException
     private $file;
     private $line;
 
-    public static function create(\Exception $exception, int $statusCode = null, array $headers = []): self
+    public static function create(\Throwable $exception, $statusCode = null, array $headers = []): self
     {
         return static::createFromThrowable($exception, $statusCode, $headers);
     }
@@ -106,7 +104,7 @@ class FlattenException
     /**
      * @return $this
      */
-    public function setStatusCode($code)
+    public function setStatusCode($code): self
     {
         $this->statusCode = $code;
 
@@ -121,7 +119,7 @@ class FlattenException
     /**
      * @return $this
      */
-    public function setHeaders(array $headers)
+    public function setHeaders(array $headers): self
     {
         $this->headers = $headers;
 
@@ -136,7 +134,7 @@ class FlattenException
     /**
      * @return $this
      */
-    public function setClass($class)
+    public function setClass($class): self
     {
         $this->class = 'c' === $class[0] && 0 === strpos($class, "class@anonymous\0") ? get_parent_class($class).'@anonymous' : $class;
 
@@ -151,7 +149,7 @@ class FlattenException
     /**
      * @return $this
      */
-    public function setFile($file)
+    public function setFile($file): self
     {
         $this->file = $file;
 
@@ -166,7 +164,7 @@ class FlattenException
     /**
      * @return $this
      */
-    public function setLine($line)
+    public function setLine($line): self
     {
         $this->line = $line;
 
@@ -193,7 +191,7 @@ class FlattenException
     /**
      * @return $this
      */
-    public function setMessage($message)
+    public function setMessage($message): self
     {
         if (false !== strpos($message, "class@anonymous\0")) {
             $message = preg_replace_callback('/class@anonymous\x00.*?\.php0x?[0-9a-fA-F]++/', function ($m) {
@@ -214,7 +212,7 @@ class FlattenException
     /**
      * @return $this
      */
-    public function setCode($code)
+    public function setCode($code): self
     {
         $this->code = $code;
 
@@ -229,7 +227,7 @@ class FlattenException
     /**
      * @return $this
      */
-    public function setPrevious(self $previous)
+    public function setPrevious(self $previous): self
     {
         $this->previous = $previous;
 
@@ -262,7 +260,7 @@ class FlattenException
     /**
      * @return $this
      */
-    public function setTrace($trace, $file, $line)
+    public function setTrace($trace, $file, $line): self
     {
         $this->trace = [];
         $this->trace[] = [